% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHRUparameters.R
\name{getHRUparameters}
\alias{getHRUparameters}
\title{Reads parameters from CRHM .prj file}
\usage{
getHRUparameters(prj_file)
}
\arguments{
\item{prj_file}{Name of prj file}
}
\value{
Returns a list containing: HRU_name, HRU_SD_max, HRU_area
}
\description{
Reads parameters from CRHM .prj file
}
\examples{
\dontrun{HRU_params <- getHRUparameters("model.prj")}
}
