% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpondsizes.R
\name{getpondsizes}
\alias{getpondsizes}
\title{Returns parabolic pond sizes: area, perimeter and depth}
\usage{
getpondsizes(CRHM_output_file, returned)
}
\arguments{
\item{CRHM_output_file}{has all the crhm stuff}

\item{returned}{extracted values for area and sd_max}
}
\value{
Pond area, perimeter and depth on the parabola using CRHMr
}
\description{
Returns parabolic pond sizes: area, perimeter and depth
}
\examples{
\dontrun
}
