% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipin.R
\name{manipin}
\alias{manipin}
\title{Processes Inflow Data from CRHM}
\usage{
manipin(pht.in)
}
\arguments{
\item{pht.in}{Name of unprocessed crhm inflow outputs}
}
\value{
temp.in that has been cleaned, reformated and ready to use to calculate water balance
}
\description{
Processes Inflow Data from CRHM
}
\examples{
\dontrun
}
