% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipoutst.R
\name{manipoutst}
\alias{manipoutst}
\title{Processes Outflow and Storage Data from CRHM}
\usage{
manipoutst(pht.outst)
}
\arguments{
\item{pht.outst}{Name of unprocessed crhm outflow and storage outputs}
}
\value{
temp.outst that has been cleaned, reformated and ready to use to calculate water balance
}
\description{
Processes Outflow and Storage Data from CRHM
}
\examples{
\dontrun
}
