% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeinout.R
\name{mergeinout}
\alias{mergeinout}
\title{merge pondsizes, tempin, tempoutst}
\usage{
mergeinout(ps.w, temp.in, temp.outst)
}
\arguments{
\item{ps.w}{pond sizes}

\item{temp.in}{crhm inflow}

\item{temp.outst}{crhm outflow}
}
\value{
combined dataframe
}
\description{
merge pondsizes, tempin, tempoutst
}
\examples{
\dontrun
}
