% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbcalc.R
\name{wbcalc}
\alias{wbcalc}
\title{Calculate inflow/outflow}
\usage{
wbcalc(df)
}
\arguments{
\item{df}{dataframe merged from pond sizes, in, out and storage dataframe}
}
\value{
f.df dataframe with calculated inflow, out, and outflow
}
\description{
Calculate inflow/outflow
}
\examples{
\dontrun
}
