% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WQCRHM-package.R
\docType{package}
\name{WQCRHM}
\alias{WQCRHM}
\alias{WQCRHM-package}
\title{Functions for CRHM water quality analyses}
\description{
\pkg{WQCRHM} is intended for water quality processing of CRHM output.
}
\references{
To cite \pkg{WQCRHM} in publications, use the command \code{citation("WQCRHM")} to get the current version of the citation.\cr
}
